Move data schema into its own file.
